Proper casts and linebreaks
Should now be compatible to clang -Weverything as of 3.6.2.
I'm aware that it's unlikely that I'll ever see a platform with int having
a larger range than uint64_t, but it's possible.  Because of that, even the
casts in pe_be64toh and pe_le64toh are necessary.

The goal: be portable *as fuck*.